requirements.txt
streamlit
st-clickable-images
pillow
streamlit-image-coordinates
opencv-python
streamlit-camera-input-live
openpyxl
matplotlib
azure-ai-vision-imageanalysis
fastapi
uvicorn
python-multipart
runpod